feat/mobile-sidbar-sync-with-desktop
