+++ apps/dashboard/app/(app)/layout.tsx
@@ -32,7 +32,7 @@ export default async function Layout({ children, breadcrumb }: LayoutProps) {
     <div className="h-[100dvh] relative flex flex-col overflow-hidden bg-background lg:flex-row">
       <UsageBanner workspace={workspace} />
 
-      <MobileSideBar className="lg:hidden" />
+      <MobileSideBar className="lg:hidden" workspace={workspace} />
       <div className="flex flex-1 overflow-hidden bg-gray-100 dark:bg-gray-950">
         <DesktopSidebar
           workspace={workspace}
